#9 refactored the location of models
