+++ attendance/lib/AttendanceDatailsPage/attendance_details_page_viewmodel.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 
 import 'package:attendance/AttendanceDatailsPage/attendance_details_page.dart';
 import 'package:flutter/material.dart';
-import 'package:attendance/model/attend_model.dart';
-import 'package:attendance/model/session_model.dart';
+import 'package:attendance/backend/attend_model.dart';
+import 'package:attendance/backend/session_model.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:barcode_scan/barcode_scan.dart';
 import 'package:firebase_auth/firebase_auth.dart';
