+++ attendance/lib/backend/scan_model.dart
