+++ attendance/lib/backend/session_model.dart
